remove EPIX_SYSTEM and EPIX_SYSTEMT related macros since they cant work properly on gcc and clang. instead, add set label and set labels function to SystemAddInfo to manually set the system label. SystemT are no longer used, and epix::into(Funcs&&...) will return SystemAddInfo

+++ epix_engine/world/sand/src/simulation.cpp
@@ -4,7 +4,7 @@
 
 #include "epix/world/sand.h"
 
-#define EPIX_WORLD_SAND_DEFAULT_CHUNK_RESET_TIME 12i32
+#define EPIX_WORLD_SAND_DEFAULT_CHUNK_RESET_TIME (int)12
 
 using namespace epix::world::sand::components;
